# [perf] Use yamllib if available