@cmungall 2024-03-13
thx!

Looks like one failing test

We also want to consider whether to incorporate ruamel.yaml into our strategy. We use it in some places as it has nice features like comment preservations (very useful for diff/patch workflows). Not sure how it compares on performance

 https://yaml.readthedocs.io/en/latest/pyyaml/